pyproject.toml
[project]
name = "rag-get-started"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "gradio>=4.44.1",
    "huggingface-hub==0.23.4",
    "jinja2<3.1.5",
    "starlette<0.40.0",
    "fastapi<0.113.0",
    "langchain==0.1.16",
    "langchain-community>=0.0.38",
    "langchain-openai>=0.1.7",
    "openai>=1.109.1",
    "pydantic>=2.5.0,<2.8.0",
    "sentence-transformers==2.5.1",
    "transformers==4.41.2",
    "wget==3.2",
    "python-dotenv>=1.2.2",
    "pypdf>=6.9.2",
    "llama-index-core>=0.10.68.post1",
    "llama-index-llms-openrouter>=0.1.4",
    "llama-index-embeddings-huggingface>=0.1.5",
    "llama-index-readers-web>=0.1.23",
    "llama-hub>=0.0.79.post1",
]
